ARKANE-851 - Fix tests
