+++ provider-ethereum/src/main/java/network/arkane/provider/EthereumAutoConfiguration.java
@@ -1,0 +1,15 @@
+package network.arkane.provider;
+
+import network.arkane.provider.opensea.OpenSeaClient;
+import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
+import org.springframework.boot.autoconfigure.http.HttpMessageConvertersAutoConfiguration;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.cloud.openfeign.FeignAutoConfiguration;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableFeignClients(basePackageClasses = {OpenSeaClient.class})
+@ImportAutoConfiguration( {FeignAutoConfiguration.class, HttpMessageConvertersAutoConfiguration.class})
+public class EthereumAutoConfiguration {
+
+}
